chore: refactor job-list
