+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-06-13T13:06:30.599Z\n"
-"PO-Revision-Date: 2019-06-13T13:06:30.599Z\n"
+"POT-Creation-Date: 2019-06-14T14:11:15.798Z\n"
+"PO-Revision-Date: 2019-06-14T14:11:15.798Z\n"
 
 msgid "Are you sure you want to discard your changes?"
 msgstr ""
