+++ src/components/Buttons/AddButton.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 import FloatingActionButton from 'material-ui/FloatingActionButton';
 import FontIcon from 'material-ui/FontIcon';
 
@@ -9,11 +10,13 @@ const style = {
 };
 
 const AddButton = () => (
-    <div style={style}>
-        <FloatingActionButton>
-            <FontIcon className="material-icons">add</FontIcon>
-        </FloatingActionButton>
-    </div>
+    <Link to={'add'}>
+        <div style={style}>
+            <FloatingActionButton>
+                <FontIcon className="material-icons">add</FontIcon>
+            </FloatingActionButton>
+        </div>
+    </Link>
 );
 
 export default AddButton;
