+++ src/components/JobList/JobList.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { AddButton } from '../Buttons';
+import Header from './Header';
+import Table from './Table';
+
+const JobList = () => (
+    <div>
+        <Header />
+        <Table />
+        <AddButton />
+    </div>
+);
+
+export default JobList;
